requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
yfinance>=0.2.0
pandas-ta==0.3.14b0
requests>=2.31.0
matplotlib>=3.8.0
pyyaml>=6.0.0
vaderSentiment>=3.3.0
textblob>=0.17.0
transformers>=4.30.0
evidently>=0.4.0
alibi-detect>=0.11.0
streamlit>=1.30.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
SQLAlchemy==2.0.41
psycopg[binary]==3.1.18
alembic==1.12.1
python-dotenv==1.0.1
streamlit
matplotlib
omegaconf
ta
openpyxl
requests
tqdm
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
psycopg2-binary
sqlalchemy
alembic
psycopg2-binary